Make PeriodicReader with own Thread the default
