+++ opentelemetry-sdk/Cargo.toml
@@ -53,6 +53,7 @@ rt-tokio-current-thread = ["tokio", "tokio-stream"]
 rt-async-std = ["async-std"]
 internal-logs = ["tracing"]
 experimental_metrics_periodic_reader_no_runtime = ["metrics"]
+experimental_metrics_periodicreader_with_async_runtime = ["metrics"]
 spec_unstable_metrics_views = ["metrics"]
 
 [[bench]]
